requirements.txt
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
Django==5.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
psycopg2-binary==2.9.10
python-telegram-bot==20.7
sniffio==1.3.1
sqlparse==0.5.3
